Feature: Implement Comprehensive Logic and Database Structure

+++ src/Services/Campaign/Core/MarGate.Campaign.Application/Handlers/Campaigns/Commands/CreateCampaign/CreateCampaignCommandRequest.cs
@@ -6,7 +6,7 @@ public class CreateCampaignCommandRequest : Command<CreateCampaignCommandRespons
 {
     public string Name { get; set; }
     public string Description { get; set; }
-    public double DiscountPercentage { get; set; }
+    public decimal DiscountPercentage { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
     public bool IsActive { get; set; }
